Refactoring (#28)
* refactoring - added common.cpp

* fix compile warnings

* refactoring - reuse waveform types and functions

* adding release build types with asserts

* add forgotten cmake script for build types

* keytap2 : reuse CC functions

* keytap : fix potential integer overflow in stdev calculation

* guess_qp2 : refactor + improve

* common : toSeconds()

* refactor : reuse readFromFile()

* refactor : frand()

* refactor : reuse calculateSimilarityMap()

* remove i32 sample support due to integer overflow issues

use i16 instead

* refactor : reuse findKeyPresses()

* keytap2-gui : fix minor bug

* remove some unused headers

* refactor : keyPresses I/O

* refactoring : add common-gui

* refactoring : reuse playback callback

* minor

+++ audio_logger.cpp
@@ -127,7 +127,7 @@ struct AudioLogger::Data {
     std::array<Frame, getBufferSize_frames(kMaxSampleRate, kMaxBufferSize_s)> buffer;
 
     int32_t nRecords = 0;
-    std::array<int32_t, kMaxRecords>  nFramesToRecord;
+    std::array<int32_t, kMaxRecords> nFramesToRecord;
     std::array<Record, kMaxRecords> records;
 
     Parameters parameters;
@@ -271,7 +271,7 @@ bool AudioLogger::addFrame(const Sample * stream) {
 
     if (data.isReady == false) return false;
 
-	if ((int) SDL_GetQueuedAudioSize(data.deviceIdIn) > 32*sizeof(float)*kSamplesPerFrame) {
+	if (SDL_GetQueuedAudioSize(data.deviceIdIn) > 32*sizeof(float)*kSamplesPerFrame) {
 		printf("Queue size: %d\n", SDL_GetQueuedAudioSize(data.deviceIdIn));
 		SDL_ClearQueuedAudio(data.deviceIdIn);
 	}
@@ -364,7 +364,7 @@ bool AudioLogger::record(float bufferSize_s, int32_t nPrevFrames) {
     if (record.size() == 0) {
         int fStart = data.bufferId - nPrevFrames;
         if (fStart < 0) fStart += data.buffer.size();
-        for (size_t i = 0; i < nPrevFrames; ++i) {
+        for (int i = 0; i < nPrevFrames; ++i) {
             record.push_back(data.buffer[(fStart + i)%data.buffer.size()]);
         }
     } else {
